requirements.txt
chromadb>=0.5.5
openai>=1.40.0
pandas>=2.1.0
numpy>=1.24.0
pypdf>=4.2.0
python-dateutil>=2.9.0
pyyaml>=6.0.1
requests>=2.32.0
beautifulsoup4>=4.12.3
scikit-learn>=1.4.0
joblib>=1.3.0
pyarrow>=12.0.0
streamlit>=1.36.0
sentence-transformers>=3.0.1
matplotlib>=3.8.0
